refactor: rename UiMembershipEmailError to UiMembershipError
